Tails Chaos 4/Tornado 1 + Plane Compatibility with DC Characters
+ EV003A (Tails After Chaos 4)
+ EV003B (Tails Tornado 1)
- Tornado 1 & 2 now show up correctly with DC Characters enabled. The pointer to the plane's action now leads to the vanilla use of the same address, which SA1Chars writes to. The original pointer is commented out if anyone needs it for whatever reason.
- Float markers for EV0032
- Int cast for a part in EV0039 where the Emeralds use Eggman's Position for their Rotation.

+++ SADX-cutscene-decompilation/Tails/ev0039_t_beforechaos4.cpp
@@ -318,16 +318,16 @@ void ev0039_t_beforechaos4(int state)
 			eggmoble->Data1->Position.y - 9.1999998f,
 			eggmoble->Data1->Position.z);
 		EV_SetAng(G_EME,
-			eggmoble->Data1->Position.x,
-			eggmoble->Data1->Position.y,
-			eggmoble->Data1->Position.z);
+			(int)eggmoble->Data1->Position.x,
+			(int)eggmoble->Data1->Position.y,
+			(int)eggmoble->Data1->Position.z);
 		EV_SetPos(B_EME, eggmoble->Data1->Position.x,
 			eggmoble->Data1->Position.y - 9.1999998f,
 			eggmoble->Data1->Position.z);
 		EV_SetAng(B_EME,
-			eggmoble->Data1->Position.x,
-			eggmoble->Data1->Position.y,
-			eggmoble->Data1->Position.z);
+			(int)eggmoble->Data1->Position.x,
+			(int)eggmoble->Data1->Position.y,
+			(int)eggmoble->Data1->Position.z);
 		EV_SetPos(sonic, 37.299999f, 72.019997f, 213.32001f);
 		EV_LookFree(player);
 		EV_ClrAction(player);
